requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.5
django-nested-admin==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==26.0.0
idna==3.7
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
